refactor(web): LlmPlaygroundPage 2287 行按缝拆五文件,修 typecheck 空转
- playground-editor.ts(~700 行):EditorMessage/EditorTool 等域类型 + payload
  解析/序列化/编辑器构造纯函数,无 JSX 可单测
- message-editor-card.tsx / tool-editor-card.tsx:两张编辑卡组件
- playground-ui.tsx:Panel/Field/StateHint/ToolCallCard/MetaItem 小件
- 主页面剩 ~1040 行(状态编排 + 布局)
纯移动零行为变化。顺手修一个真 bug:apps/web 的 typecheck 脚本对
solution-style tsconfig(files:[])跑 tsc --noEmit 实际检查 0 个文件,一直
形同虚设;改为 tsc -b(build mode 检查全部 references)。
