+++ apps/web/src/pages/llm-playground/LlmPlaygroundPage.tsx
@@ -1,27 +1,15 @@
 import {
-  LlmPlaygroundChatRequestSchema,
   LlmPlaygroundChatResponseSchema,
   LlmPlaygroundToolListResponseSchema,
   LlmProviderListResponseSchema,
   type LlmPlaygroundChatRequest,
   type LlmPlaygroundChatResponse,
-  type LlmToolCallPayload,
   type LlmToolDefinition,
   type LlmProviderOption,
-  type PlaygroundContentPart,
-  type PlaygroundMessage,
 } from "@kagami/shared/schemas/llm-chat";
 import { useMutation, useQuery } from "@tanstack/react-query";
-import {
-  ArrowDown,
-  ArrowUp,
-  ImagePlus,
-  Plus,
-  RefreshCcw,
-  SendHorizontal,
-  Trash2,
-} from "lucide-react";
-import { useEffect, useEffectEvent, useMemo, useRef, useState, type ReactNode } from "react";
+import { Plus, RefreshCcw, SendHorizontal } from "lucide-react";
+import { useEffect, useEffectEvent, useMemo, useRef, useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import { Button } from "@/components/ui/button";
 import {
@@ -42,77 +30,34 @@ import {
   getPlaygroundImportDraftFromLocationState,
   resolvePlaygroundImportDraft,
 } from "./playground-import";
-
-type PlaygroundResult = {
-  payload: LlmPlaygroundChatRequest;
-  response: ApiRequestResult;
-  parsedResponse: LlmPlaygroundChatResponse | null;
-  responseSchemaError: string | null;
-};
-
-type ToolChoiceMode = "none" | "auto" | "required" | "tool";
-type EditorRole = PlaygroundMessage["role"];
-type EditorPropertyType = "string" | "number" | "integer" | "boolean" | "object" | "array";
-
-type EditorTextPart = {
-  id: string;
-  type: "text";
-  text: string;
-};
-
-type EditorImagePart = {
-  id: string;
-  type: "image";
-  fileName: string;
-  mimeType: string;
-  dataUrl: string;
-};
-
-type EditorContentPart = EditorTextPart | EditorImagePart;
-
-type EditorToolCall = {
-  id: string;
-  toolCallId: string;
-  name: string;
-  argumentsText: string;
-};
-
-type EditorUserMessage = {
-  id: string;
-  role: "user";
-  parts: EditorContentPart[];
-};
-
-type EditorAssistantMessage = {
-  id: string;
-  role: "assistant";
-  content: string;
-  toolCalls: EditorToolCall[];
-};
-
-type EditorToolMessage = {
-  id: string;
-  role: "tool";
-  toolCallId: string;
-  content: string;
-};
-
-type EditorMessage = EditorUserMessage | EditorAssistantMessage | EditorToolMessage;
-
-type EditorToolProperty = {
-  id: string;
-  name: string;
-  type: EditorPropertyType;
-  description: string;
-  rawSchema: Record<string, unknown>;
-};
-
-type EditorTool = {
-  id: string;
-  name: string;
-  description: string;
-  properties: EditorToolProperty[];
-};
+import {
+  convertMessageRole,
+  createDefaultMessages,
+  createEditorId,
+  createMessage,
+  createTextPart,
+  createToolCall,
+  createToolEditor,
+  createToolProperty,
+  formatHttpError,
+  formatSchemaIssues,
+  moveItem,
+  parsePlaygroundPayload,
+  playgroundMessageToEditor,
+  readFileAsDataUrl,
+  toolDefinitionToEditor,
+  type EditorContentPart,
+  type EditorMessage,
+  type EditorRole,
+  type EditorTool,
+  type EditorToolCall,
+  type EditorToolProperty,
+  type PlaygroundResult,
+  type ToolChoiceMode,
+} from "./playground-editor";
+import { MessageEditorCard } from "./message-editor-card";
+import { ToolEditorCard } from "./tool-editor-card";
+import { Field, MetaItem, Panel, StateHint, ToolCallCard } from "./playground-ui";
 
 const EMPTY_PROVIDERS: LlmProviderOption[] = [];
 const EMPTY_TOOL_LIBRARY: LlmToolDefinition[] = [];
@@ -1068,1220 +1013,6 @@ export function LlmPlaygroundPage() {
           </div>
         </div>
       </div>
-    </div>
-  );
-}
-
-function MessageEditorCard({
-  index,
-  message,
-  total,
-  onDelete,
-  onMoveUp,
-  onMoveDown,
-  onRoleChange,
-  onTextChange,
-  onToolCallIdChange,
-  onAddTextPart,
-  onAddImagePart,
-  onPartChange,
-  onPartMove,
-  onPartDelete,
-  onAddAssistantToolCall,
-  onAssistantToolCallChange,
-  onAssistantToolCallDelete,
-}: {
-  index: number;
-  message: EditorMessage;
-  total: number;
-  onDelete: () => void;
-  onMoveUp: () => void;
-  onMoveDown: () => void;
-  onRoleChange: (role: EditorRole) => void;
-  onTextChange: (value: string) => void;
-  onToolCallIdChange: (value: string) => void;
-  onAddTextPart: () => void;
-  onAddImagePart: (file: File | null) => void;
-  onPartChange: (partId: string, updater: (part: EditorContentPart) => EditorContentPart) => void;
-  onPartMove: (partId: string, direction: "up" | "down") => void;
-  onPartDelete: (partId: string) => void;
-  onAddAssistantToolCall: () => void;
-  onAssistantToolCallChange: (
-    toolCallId: string,
-    updater: (toolCall: EditorToolCall) => EditorToolCall,
-  ) => void;
-  onAssistantToolCallDelete: (toolCallId: string) => void;
-}) {
-  return (
-    <section className="rounded-none border bg-background/80 p-4">
-      <div className="flex flex-wrap items-start justify-between gap-3">
-        <div className="space-y-1">
-          <p className="text-xs uppercase tracking-wide text-muted-foreground">
-            Message {index + 1}
-          </p>
-          <div className="w-[180px]">
-            <Select value={message.role} onValueChange={value => onRoleChange(value as EditorRole)}>
-              <SelectTrigger aria-label={`Message ${index + 1} role`}>
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="user">user</SelectItem>
-                <SelectItem value="assistant">assistant</SelectItem>
-                <SelectItem value="tool">tool</SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
-        </div>
-
-        <div className="flex gap-2">
-          <Button
-            type="button"
-            size="icon"
-            variant="outline"
-            onClick={onMoveUp}
-            disabled={index === 0}
-          >
-            <ArrowUp className="h-4 w-4" />
-          </Button>
-          <Button
-            type="button"
-            size="icon"
-            variant="outline"
-            onClick={onMoveDown}
-            disabled={index === total - 1}
-          >
-            <ArrowDown className="h-4 w-4" />
-          </Button>
-          <Button type="button" size="icon" variant="outline" onClick={onDelete}>
-            <Trash2 className="h-4 w-4" />
-          </Button>
-        </div>
-      </div>
-
-      {message.role === "user" ? (
-        <div className="mt-4 space-y-3">
-          <div className="flex flex-wrap gap-2">
-            <Button type="button" size="sm" onClick={onAddTextPart}>
-              <Plus className="mr-2 h-4 w-4" />
-              文本片段
-            </Button>
-            <label className="inline-flex cursor-pointer items-center justify-center whitespace-nowrap rounded-none border border-input bg-background px-3 py-2 text-sm font-medium hover:bg-accent hover:text-accent-foreground">
-              <ImagePlus className="mr-2 h-4 w-4" />
-              上传图片
-              <input
-                type="file"
-                accept="image/*"
-                className="hidden"
-                onChange={event => {
-                  const file = event.target.files?.[0] ?? null;
-                  void onAddImagePart(file);
-                  event.target.value = "";
-                }}
-              />
-            </label>
-          </div>
-
-          {message.parts.length === 0 ? (
-            <StateHint text="这条 user 消息还没有内容。" />
-          ) : (
-            <div className="space-y-3">
-              {message.parts.map((part, partIndex) => (
-                <section key={part.id} className="rounded-none border bg-muted/20 p-3">
-                  <div className="mb-3 flex items-center justify-between gap-3">
-                    <div>
-                      <p className="text-sm font-medium">
-                        {part.type === "text" ? "文本片段" : "图片片段"}
-                      </p>
-                      <p className="text-xs text-muted-foreground">片段 {partIndex + 1}</p>
-                    </div>
-                    <div className="flex gap-2">
-                      <Button
-                        type="button"
-                        size="icon"
-                        variant="outline"
-                        onClick={() => onPartMove(part.id, "up")}
-                        disabled={partIndex === 0}
-                      >
-                        <ArrowUp className="h-4 w-4" />
-                      </Button>
-                      <Button
-                        type="button"
-                        size="icon"
-                        variant="outline"
-                        onClick={() => onPartMove(part.id, "down")}
-                        disabled={partIndex === message.parts.length - 1}
-                      >
-                        <ArrowDown className="h-4 w-4" />
-                      </Button>
-                      <Button
-                        type="button"
-                        size="icon"
-                        variant="outline"
-                        onClick={() => onPartDelete(part.id)}
-                      >
-                        <Trash2 className="h-4 w-4" />
-                      </Button>
-                    </div>
-                  </div>
-
-                  {part.type === "text" ? (
-                    <textarea
-                      value={part.text}
-                      onChange={event =>
-                        onPartChange(part.id, currentPart =>
-                          currentPart.type === "text"
-                            ? {
-                                ...currentPart,
-                                text: event.target.value,
-                              }
-                            : currentPart,
-                        )
-                      }
-                      placeholder="输入文本内容"
-                      className="min-h-[110px] w-full resize-y rounded-none border bg-background p-3 text-sm outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-                    />
-                  ) : (
-                    <div className="grid gap-3 md:grid-cols-[180px_minmax(0,1fr)]">
-                      <img
-                        src={part.dataUrl}
-                        alt={part.fileName || "uploaded"}
-                        className="h-36 w-full rounded-none border object-cover"
-                      />
-                      <div className="space-y-3">
-                        <Field label="文件名">
-                          <input
-                            value={part.fileName}
-                            onChange={event =>
-                              onPartChange(part.id, currentPart =>
-                                currentPart.type === "image"
-                                  ? {
-                                      ...currentPart,
-                                      fileName: event.target.value,
-                                    }
-                                  : currentPart,
-                              )
-                            }
-                            className="h-10 w-full rounded-none border bg-background px-3 text-sm outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-                          />
-                        </Field>
-                        <Field label="MIME Type">
-                          <input
-                            value={part.mimeType}
-                            onChange={event =>
-                              onPartChange(part.id, currentPart =>
-                                currentPart.type === "image"
-                                  ? {
-                                      ...currentPart,
-                                      mimeType: event.target.value,
-                                    }
-                                  : currentPart,
-                              )
-                            }
-                            className="h-10 w-full rounded-none border bg-background px-3 text-sm outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-                          />
-                        </Field>
-                      </div>
-                    </div>
-                  )}
-                </section>
-              ))}
-            </div>
-          )}
-        </div>
-      ) : null}
-
-      {message.role === "assistant" ? (
-        <div className="mt-4 space-y-4">
-          <textarea
-            value={message.content}
-            onChange={event => onTextChange(event.target.value)}
-            placeholder="assistant 文本内容"
-            className="min-h-[120px] w-full resize-y rounded-none border bg-background p-3 text-sm outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-          />
-
-          <div className="rounded-none border border-dashed bg-muted/20 p-4">
-            <div className="flex flex-wrap items-center justify-between gap-3">
-              <div>
-                <h3 className="text-sm font-semibold">Tool Calls</h3>
-                <p className="text-xs text-muted-foreground">可选，用于构造 assistant 历史响应。</p>
-              </div>
-              <Button type="button" size="sm" variant="outline" onClick={onAddAssistantToolCall}>
-                <Plus className="mr-2 h-4 w-4" />
-                新增 Tool Call
-              </Button>
-            </div>
-
-            {message.toolCalls.length === 0 ? (
-              <p className="mt-3 text-sm text-muted-foreground">当前没有 tool calls。</p>
-            ) : (
-              <div className="mt-3 space-y-3">
-                {message.toolCalls.map(toolCall => (
-                  <section key={toolCall.id} className="rounded-none border bg-background p-3">
-                    <div className="mb-3 flex justify-end">
-                      <Button
-                        type="button"
-                        size="icon"
-                        variant="outline"
-                        onClick={() => onAssistantToolCallDelete(toolCall.id)}
-                      >
-                        <Trash2 className="h-4 w-4" />
-                      </Button>
-                    </div>
-
-                    <div className="grid gap-3 md:grid-cols-2">
-                      <Field label="Tool Call ID">
-                        <input
-                          value={toolCall.toolCallId}
-                          onChange={event =>
-                            onAssistantToolCallChange(toolCall.id, currentToolCall => ({
-                              ...currentToolCall,
-                              toolCallId: event.target.value,
-                            }))
-                          }
-                          className="h-10 w-full rounded-none border bg-background px-3 text-sm outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-                        />
-                      </Field>
-                      <Field label="Tool Name">
-                        <input
-                          value={toolCall.name}
-                          onChange={event =>
-                            onAssistantToolCallChange(toolCall.id, currentToolCall => ({
-                              ...currentToolCall,
-                              name: event.target.value,
-                            }))
-                          }
-                          className="h-10 w-full rounded-none border bg-background px-3 text-sm outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-                        />
-                      </Field>
-                    </div>
-
-                    <Field label="Arguments JSON" className="mt-3">
-                      <textarea
-                        value={toolCall.argumentsText}
-                        onChange={event =>
-                          onAssistantToolCallChange(toolCall.id, currentToolCall => ({
-                            ...currentToolCall,
-                            argumentsText: event.target.value,
-                          }))
-                        }
-                        className="min-h-[110px] w-full resize-y rounded-none border bg-background p-3 font-mono text-xs outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-                      />
-                    </Field>
-                  </section>
-                ))}
-              </div>
-            )}
-          </div>
-        </div>
-      ) : null}
-
-      {message.role === "tool" ? (
-        <div className="mt-4 space-y-3">
-          <Field label="Tool Call ID">
-            <input
-              value={message.toolCallId}
-              onChange={event => onToolCallIdChange(event.target.value)}
-              placeholder="例如 call_123"
-              className="h-10 w-full rounded-none border bg-background px-3 text-sm outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-            />
-          </Field>
-          <Field label="Tool Output">
-            <textarea
-              value={message.content}
-              onChange={event => onTextChange(event.target.value)}
-              placeholder="工具返回内容"
-              className="min-h-[120px] w-full resize-y rounded-none border bg-background p-3 text-sm outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-            />
-          </Field>
-        </div>
-      ) : null}
-    </section>
-  );
-}
-
-function ToolEditorCard({
-  tool,
-  index,
-  total,
-  onMoveUp,
-  onMoveDown,
-  onDelete,
-  onChange,
-  onAddProperty,
-  onPropertyChange,
-  onPropertyMove,
-  onPropertyDelete,
-}: {
-  tool: EditorTool;
-  index: number;
-  total: number;
-  onMoveUp: () => void;
-  onMoveDown: () => void;
-  onDelete: () => void;
-  onChange: (updater: (tool: EditorTool) => EditorTool) => void;
-  onAddProperty: () => void;
-  onPropertyChange: (
-    propertyId: string,
-    updater: (property: EditorToolProperty) => EditorToolProperty,
-  ) => void;
-  onPropertyMove: (propertyId: string, direction: "up" | "down") => void;
-  onPropertyDelete: (propertyId: string) => void;
-}) {
-  return (
-    <section className="rounded-none border bg-background/80 p-4">
-      <div className="flex flex-wrap items-start justify-between gap-3">
-        <div className="space-y-1">
-          <p className="text-xs uppercase tracking-wide text-muted-foreground">Tool {index + 1}</p>
-          <p className="text-sm font-semibold">{tool.name.trim() || "未命名工具"}</p>
-        </div>
-
-        <div className="flex gap-2">
-          <Button
-            type="button"
-            size="icon"
-            variant="outline"
-            onClick={onMoveUp}
-            disabled={index === 0}
-          >
-            <ArrowUp className="h-4 w-4" />
-          </Button>
-          <Button
-            type="button"
-            size="icon"
-            variant="outline"
-            onClick={onMoveDown}
-            disabled={index === total - 1}
-          >
-            <ArrowDown className="h-4 w-4" />
-          </Button>
-          <Button type="button" size="icon" variant="outline" onClick={onDelete}>
-            <Trash2 className="h-4 w-4" />
-          </Button>
-        </div>
-      </div>
-
-      <div className="mt-4 grid gap-3 md:grid-cols-2">
-        <Field label="Name">
-          <input
-            value={tool.name}
-            onChange={event =>
-              onChange(currentTool => ({
-                ...currentTool,
-                name: event.target.value,
-              }))
-            }
-            className="h-10 w-full rounded-none border bg-background px-3 text-sm outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-          />
-        </Field>
-        <Field label="Description">
-          <input
-            value={tool.description}
-            onChange={event =>
-              onChange(currentTool => ({
-                ...currentTool,
-                description: event.target.value,
-              }))
-            }
-            className="h-10 w-full rounded-none border bg-background px-3 text-sm outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-          />
-        </Field>
-      </div>
-
-      <div className="mt-4 rounded-none border border-dashed bg-muted/20 p-4">
-        <div className="flex flex-wrap items-center justify-between gap-3">
-          <div>
-            <h3 className="text-sm font-semibold">Properties</h3>
-            <p className="text-xs text-muted-foreground">`parameters.type` 固定为 object。</p>
-          </div>
-          <Button type="button" size="sm" variant="outline" onClick={onAddProperty}>
-            <Plus className="mr-2 h-4 w-4" />
-            新增属性
-          </Button>
-        </div>
-
-        {tool.properties.length === 0 ? (
-          <p className="mt-3 text-sm text-muted-foreground">当前没有参数属性。</p>
-        ) : (
-          <div className="mt-3 space-y-3">
-            {tool.properties.map((property, propertyIndex) => (
-              <section key={property.id} className="rounded-none border bg-background p-3">
-                <div className="mb-3 flex items-center justify-between gap-3">
-                  <div>
-                    <p className="text-sm font-medium">{property.name.trim() || "未命名属性"}</p>
-                    <p className="text-xs text-muted-foreground">属性 {propertyIndex + 1}</p>
-                  </div>
-                  <div className="flex gap-2">
-                    <Button
-                      type="button"
-                      size="icon"
-                      variant="outline"
-                      onClick={() => onPropertyMove(property.id, "up")}
-                      disabled={propertyIndex === 0}
-                    >
-                      <ArrowUp className="h-4 w-4" />
-                    </Button>
-                    <Button
-                      type="button"
-                      size="icon"
-                      variant="outline"
-                      onClick={() => onPropertyMove(property.id, "down")}
-                      disabled={propertyIndex === tool.properties.length - 1}
-                    >
-                      <ArrowDown className="h-4 w-4" />
-                    </Button>
-                    <Button
-                      type="button"
-                      size="icon"
-                      variant="outline"
-                      onClick={() => onPropertyDelete(property.id)}
-                    >
-                      <Trash2 className="h-4 w-4" />
-                    </Button>
-                  </div>
-                </div>
-
-                <div className="grid gap-3 md:grid-cols-[minmax(0,1fr)_160px]">
-                  <Field label="Property Name">
-                    <input
-                      value={property.name}
-                      onChange={event =>
-                        onPropertyChange(property.id, currentProperty => ({
-                          ...currentProperty,
-                          name: event.target.value,
-                        }))
-                      }
-                      className="h-10 w-full rounded-none border bg-background px-3 text-sm outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-                    />
-                  </Field>
-
-                  <Field label="Type">
-                    <Select
-                      value={property.type}
-                      onValueChange={value =>
-                        onPropertyChange(property.id, currentProperty => ({
-                          ...currentProperty,
-                          type: value as EditorPropertyType,
-                          rawSchema: {
-                            ...currentProperty.rawSchema,
-                            type: value,
-                          },
-                        }))
-                      }
-                    >
-                      <SelectTrigger aria-label="Property Type">
-                        <SelectValue />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="string">string</SelectItem>
-                        <SelectItem value="number">number</SelectItem>
-                        <SelectItem value="integer">integer</SelectItem>
-                        <SelectItem value="boolean">boolean</SelectItem>
-                        <SelectItem value="object">object</SelectItem>
-                        <SelectItem value="array">array</SelectItem>
-                      </SelectContent>
-                    </Select>
-                  </Field>
-                </div>
-
-                <Field label="Description" className="mt-3">
-                  <input
-                    value={property.description}
-                    onChange={event =>
-                      onPropertyChange(property.id, currentProperty => ({
-                        ...currentProperty,
-                        description: event.target.value,
-                      }))
-                    }
-                    className="h-10 w-full rounded-none border bg-background px-3 text-sm outline-none ring-offset-background transition focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-                  />
-                </Field>
-              </section>
-            ))}
-          </div>
-        )}
-      </div>
-    </section>
-  );
-}
-
-function parsePlaygroundPayload({
-  provider,
-  model,
-  availableModels,
-  systemPrompt,
-  messages,
-  tools,
-  toolChoiceMode,
-  selectedToolNameForChoice,
-}: {
-  provider: LlmProviderOption["id"];
-  model: string;
-  availableModels: string[];
-  systemPrompt: string;
-  messages: EditorMessage[];
-  tools: EditorTool[];
-  toolChoiceMode: ToolChoiceMode;
-  selectedToolNameForChoice: string;
-}):
-  | {
-      success: true;
-      data: LlmPlaygroundChatRequest;
-    }
-  | {
-      success: false;
-      error: string;
-    } {
-  if (model.trim().length === 0) {
-    return {
-      success: false,
-      error: "请先选择 model。",
-    };
-  }
-
-  if (!availableModels.includes(model)) {
-    return {
-      success: false,
-      error: "所选 model 不在当前 provider 的配置列表中。",
-    };
-  }
-
-  const parsedMessages = serializeMessages(messages);
-  if (!parsedMessages.success) {
-    return parsedMessages;
-  }
-
-  const parsedTools = serializeTools(tools);
-  if (!parsedTools.success) {
-    return parsedTools;
-  }
-
-  const toolChoice = serializeToolChoice({
-    toolChoiceMode,
-    selectedToolNameForChoice,
-    currentToolNames: parsedTools.data.map(tool => tool.name),
-  });
-  if (!toolChoice.success) {
-    return toolChoice;
-  }
-
-  const payloadParsed = LlmPlaygroundChatRequestSchema.safeParse({
-    provider,
-    model,
-    ...(systemPrompt.trim().length > 0 ? { system: systemPrompt.trim() } : {}),
-    messages: parsedMessages.data,
-    tools: parsedTools.data,
-    toolChoice: toolChoice.data,
-  });
-
-  if (!payloadParsed.success) {
-    return {
-      success: false,
-      error: `提交参数校验失败：${formatSchemaIssues(payloadParsed.error.issues)}`,
-    };
-  }
-
-  return {
-    success: true,
-    data: payloadParsed.data,
-  };
-}
-
-function serializeMessages(messages: EditorMessage[]):
-  | {
-      success: true;
-      data: PlaygroundMessage[];
-    }
-  | {
-      success: false;
-      error: string;
-    } {
-  const serializedMessages: PlaygroundMessage[] = [];
-
-  for (const [index, message] of messages.entries()) {
-    if (message.role === "user") {
-      const serializedParts = message.parts
-        .map(part => {
-          if (part.type === "text") {
-            const text = part.text.trim();
-            if (text.length === 0) {
-              return null;
-            }
-
-            return {
-              type: "text",
-              text,
-            } as const;
-          }
-
-          if (part.dataUrl.trim().length === 0) {
-            return null;
-          }
-
-          return {
-            type: "image",
-            fileName: part.fileName.trim() || undefined,
-            mimeType: part.mimeType.trim(),
-            dataUrl: part.dataUrl,
-          } as const;
-        })
-        .filter((part): part is PlaygroundContentPart => part !== null);
-
-      if (serializedParts.length === 0) {
-        return {
-          success: false,
-          error: `第 ${index + 1} 条 user 消息没有有效内容。`,
-        };
-      }
-
-      if (
-        serializedParts.length === 1 &&
-        serializedParts[0].type === "text" &&
-        serializedParts[0].text.trim().length > 0
-      ) {
-        serializedMessages.push({
-          role: "user",
-          content: serializedParts[0].text,
-        });
-      } else {
-        serializedMessages.push({
-          role: "user",
-          content: serializedParts,
-        });
-      }
-
-      continue;
-    }
-
-    if (message.role === "assistant") {
-      const toolCalls: LlmToolCallPayload[] = [];
-      for (const toolCall of message.toolCalls) {
-        const toolCallId = toolCall.toolCallId.trim();
-        const toolName = toolCall.name.trim();
-        if (
-          toolCallId.length === 0 &&
-          toolName.length === 0 &&
-          toolCall.argumentsText.trim().length === 0
-        ) {
-          continue;
-        }
-
-        let parsedArguments: unknown;
-        try {
-          parsedArguments = JSON.parse(toolCall.argumentsText || "{}");
-        } catch (error) {
-          return {
-            success: false,
-            error: `第 ${index + 1} 条 assistant 消息的 tool call 参数 JSON 解析失败：${
-              error instanceof Error ? error.message : "未知错误"
-            }`,
-          };
-        }
-
-        if (
-          !parsedArguments ||
-          typeof parsedArguments !== "object" ||
-          Array.isArray(parsedArguments)
-        ) {
-          return {
-            success: false,
-            error: `第 ${index + 1} 条 assistant 消息的 tool call 参数必须是 JSON object。`,
-          };
-        }
-
-        toolCalls.push({
-          id: toolCallId,
-          name: toolName,
-          arguments: parsedArguments as Record<string, unknown>,
-        });
-      }
-
-      serializedMessages.push({
-        role: "assistant",
-        content: message.content,
-        toolCalls,
-      });
-      continue;
-    }
-
-    serializedMessages.push({
-      role: "tool",
-      toolCallId: message.toolCallId,
-      content: message.content,
-    });
-  }
-
-  if (serializedMessages.length === 0) {
-    return {
-      success: false,
-      error: "至少需要一条消息。",
-    };
-  }
-
-  return {
-    success: true,
-    data: serializedMessages,
-  };
-}
-
-function serializeTools(tools: EditorTool[]):
-  | {
-      success: true;
-      data: LlmToolDefinition[];
-    }
-  | {
-      success: false;
-      error: string;
-    } {
-  const serializedTools: LlmToolDefinition[] = [];
-  const seenToolNames = new Set<string>();
-
-  for (const [index, tool] of tools.entries()) {
-    const toolName = tool.name.trim();
-    if (toolName.length === 0) {
-      return {
-        success: false,
-        error: `第 ${index + 1} 个工具缺少名称。`,
-      };
-    }
-
-    if (seenToolNames.has(toolName)) {
-      return {
-        success: false,
-        error: `工具名称重复：${toolName}`,
-      };
-    }
-    seenToolNames.add(toolName);
-
-    const properties: Record<string, unknown> = {};
-    const seenPropertyNames = new Set<string>();
-    for (const property of tool.properties) {
-      const propertyName = property.name.trim();
-      if (propertyName.length === 0) {
-        return {
-          success: false,
-          error: `工具 ${toolName} 存在未命名属性。`,
-        };
-      }
-
-      if (seenPropertyNames.has(propertyName)) {
-        return {
-          success: false,
-          error: `工具 ${toolName} 的属性名重复：${propertyName}`,
-        };
-      }
-      seenPropertyNames.add(propertyName);
-
-      const nextSchema: Record<string, unknown> = {
-        ...property.rawSchema,
-        type: property.type,
-      };
-      if (property.description.trim().length > 0) {
-        nextSchema.description = property.description.trim();
-      } else {
-        delete nextSchema.description;
-      }
-
-      properties[propertyName] = nextSchema;
-    }
-
-    serializedTools.push({
-      name: toolName,
-      ...(tool.description.trim().length > 0 ? { description: tool.description.trim() } : {}),
-      parameters: {
-        type: "object",
-        properties,
-      },
-    });
-  }
-
-  return {
-    success: true,
-    data: serializedTools,
-  };
-}
-
-function serializeToolChoice({
-  toolChoiceMode,
-  selectedToolNameForChoice,
-  currentToolNames,
-}:
-  | {
-      toolChoiceMode: Exclude<ToolChoiceMode, "tool">;
-      selectedToolNameForChoice: string;
-      currentToolNames: string[];
-    }
-  | {
-      toolChoiceMode: "tool";
-      selectedToolNameForChoice: string;
-      currentToolNames: string[];
-    }):
-  | {
-      success: true;
-      data: LlmPlaygroundChatRequest["toolChoice"];
-    }
-  | {
-      success: false;
-      error: string;
-    } {
-  if (toolChoiceMode !== "tool") {
-    return {
-      success: true,
-      data: toolChoiceMode,
-    };
-  }
-
-  const toolName = selectedToolNameForChoice.trim();
-  if (toolName.length === 0) {
-    return {
-      success: false,
-      error: "请选择 toolChoice 对应的工具。",
-    };
-  }
-
-  if (!currentToolNames.includes(toolName)) {
-    return {
-      success: false,
-      error: "toolChoice 指定的工具必须存在于当前工具定义列表中。",
-    };
-  }
-
-  return {
-    success: true,
-    data: {
-      tool_name: toolName,
-    },
-  };
-}
-
-function formatSchemaIssues(
-  issues: Array<{
-    path: Array<string | number>;
-    message: string;
-  }>,
-): string {
-  return issues
-    .map(issue => {
-      const path = issue.path.length > 0 ? issue.path.join(".") : "(root)";
-      return `${path}: ${issue.message}`;
-    })
-    .join("; ");
-}
-
-function formatHttpError(response: ApiRequestResult): string {
-  const body =
-    typeof response.body === "string"
-      ? response.body
-      : response.body
-        ? formatJson(response.body)
-        : "";
-  return [`HTTP ${response.status} ${response.statusText}`.trim(), body].filter(Boolean).join("\n");
-}
-
-function formatJson(value: unknown): string {
-  const formatted = JSON.stringify(value, null, 2);
-  return formatted ?? "null";
-}
-
-function createDefaultMessages(): EditorMessage[] {
-  return [
-    {
-      id: createEditorId(),
-      role: "user",
-      parts: [createTextPart("你好")],
-    },
-  ];
-}
-
-function createMessage(role: EditorRole): EditorMessage {
-  if (role === "user") {
-    return {
-      id: createEditorId(),
-      role: "user",
-      parts: [createTextPart("")],
-    };
-  }
-
-  if (role === "assistant") {
-    return {
-      id: createEditorId(),
-      role: "assistant",
-      content: "",
-      toolCalls: [],
-    };
-  }
-
-  return {
-    id: createEditorId(),
-    role: "tool",
-    toolCallId: "",
-    content: "",
-  };
-}
-
-function createTextPart(text: string): EditorTextPart {
-  return {
-    id: createEditorId(),
-    type: "text",
-    text,
-  };
-}
-
-function createToolCall(): EditorToolCall {
-  return {
-    id: createEditorId(),
-    toolCallId: "",
-    name: "",
-    argumentsText: "{}",
-  };
-}
-
-function createToolEditor(): EditorTool {
-  return {
-    id: createEditorId(),
-    name: "",
-    description: "",
-    properties: [],
-  };
-}
-
-function createToolProperty(): EditorToolProperty {
-  return {
-    id: createEditorId(),
-    name: "",
-    type: "string",
-    description: "",
-    rawSchema: {
-      type: "string",
-    },
-  };
-}
-
-function toolDefinitionToEditor(tool: LlmToolDefinition): EditorTool {
-  return {
-    id: createEditorId(),
-    name: tool.name,
-    description: tool.description ?? "",
-    properties: Object.entries(tool.parameters.properties).map(([name, rawSchema]) => {
-      const objectSchema =
-        rawSchema && typeof rawSchema === "object" && !Array.isArray(rawSchema)
-          ? (rawSchema as Record<string, unknown>)
-          : {};
-
-      return {
-        id: createEditorId(),
-        name,
-        type: normalizeEditorPropertyType(objectSchema.type),
-        description: typeof objectSchema.description === "string" ? objectSchema.description : "",
-        rawSchema: objectSchema,
-      };
-    }),
-  };
-}
-
-function playgroundMessageToEditor(message: PlaygroundMessage): EditorMessage {
-  if (message.role === "user") {
-    if (typeof message.content === "string") {
-      return {
-        id: createEditorId(),
-        role: "user",
-        parts: [createTextPart(message.content)],
-      };
-    }
-
-    return {
-      id: createEditorId(),
-      role: "user",
-      parts: message.content.map(playgroundContentPartToEditor),
-    };
-  }
-
-  if (message.role === "assistant") {
-    return {
-      id: createEditorId(),
-      role: "assistant",
-      content: message.content,
-      toolCalls: message.toolCalls.map(toolCall => ({
-        id: createEditorId(),
-        toolCallId: toolCall.id,
-        name: toolCall.name,
-        argumentsText: formatJson(toolCall.arguments),
-      })),
-    };
-  }
-
-  return {
-    id: createEditorId(),
-    role: "tool",
-    toolCallId: message.toolCallId,
-    content: message.content,
-  };
-}
-
-function playgroundContentPartToEditor(part: PlaygroundContentPart): EditorContentPart {
-  if (part.type === "text") {
-    return createTextPart(part.text);
-  }
-
-  return {
-    id: createEditorId(),
-    type: "image",
-    fileName: part.fileName ?? "",
-    mimeType: part.mimeType,
-    dataUrl: part.dataUrl,
-  };
-}
-
-function normalizeEditorPropertyType(value: unknown): EditorPropertyType {
-  if (
-    value === "string" ||
-    value === "number" ||
-    value === "integer" ||
-    value === "boolean" ||
-    value === "object" ||
-    value === "array"
-  ) {
-    return value;
-  }
-
-  return "string";
-}
-
-function convertMessageRole(message: EditorMessage, role: EditorRole): EditorMessage {
-  if (message.role === role) {
-    return message;
-  }
-
-  return createMessage(role);
-}
-
-function moveItem<T extends { id: string }>(
-  items: T[],
-  itemId: string,
-  direction: "up" | "down",
-): T[] {
-  const index = items.findIndex(item => item.id === itemId);
-  if (index < 0) {
-    return items;
-  }
-
-  const nextIndex = direction === "up" ? index - 1 : index + 1;
-  if (nextIndex < 0 || nextIndex >= items.length) {
-    return items;
-  }
-
-  const nextItems = [...items];
-  const [item] = nextItems.splice(index, 1);
-  nextItems.splice(nextIndex, 0, item);
-  return nextItems;
-}
-
-function createEditorId(): string {
-  return typeof crypto !== "undefined" && typeof crypto.randomUUID === "function"
-    ? crypto.randomUUID()
-    : `id-${Math.random().toString(36).slice(2, 10)}`;
-}
-
-async function readFileAsDataUrl(file: File): Promise<string> {
-  return await new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onerror = () => {
-      reject(new Error("图片读取失败"));
-    };
-    reader.onload = () => {
-      if (typeof reader.result === "string") {
-        resolve(reader.result);
-        return;
-      }
-
-      reject(new Error("图片读取失败"));
-    };
-    reader.readAsDataURL(file);
-  });
-}
-
-function Panel({
-  title,
-  description,
-  className,
-  bodyClassName,
-  children,
-}: {
-  title: string;
-  description: string;
-  className?: string;
-  bodyClassName?: string;
-  children: ReactNode;
-}) {
-  return (
-    <section className={`rounded-none border bg-card p-5 ${className ?? ""}`}>
-      <div className="mb-4 space-y-1">
-        <h2 className="text-base font-semibold">{title}</h2>
-        <p className="text-sm text-muted-foreground">{description}</p>
-      </div>
-      <div className={bodyClassName}>{children}</div>
-    </section>
-  );
-}
-
-function Field({
-  label,
-  className,
-  children,
-}: {
-  label: string;
-  className?: string;
-  children: ReactNode;
-}) {
-  return (
-    <label className={`block space-y-2 ${className ?? ""}`}>
-      <span className="text-sm font-medium">{label}</span>
-      {children}
-    </label>
-  );
-}
-
-function StateHint({
-  text,
-  tone = "default",
-}: {
-  text: string;
-  tone?: "default" | "warning" | "error";
-}) {
-  const toneClassName =
-    tone === "error"
-      ? "border-foreground bg-signal text-signal-foreground"
-      : tone === "warning"
-        ? "border-foreground bg-scheduler text-scheduler-foreground"
-        : "border-dashed bg-muted/20 text-muted-foreground";
-
-  return (
-    <div
-      className={`flex min-h-[100px] items-center justify-center rounded-none border px-4 py-6 text-center text-sm ${toneClassName}`}
-    >
-      {text}
-    </div>
-  );
-}
-
-function ToolCallCard({ toolCall }: { toolCall: LlmToolCallPayload }) {
-  return (
-    <details className="rounded-none border bg-muted/20 p-3">
-      <summary className="cursor-pointer text-sm font-medium">
-        {toolCall.name}
-        <span className="ml-2 font-mono text-xs text-muted-foreground">{toolCall.id}</span>
-      </summary>
-      <pre className="mt-3 overflow-x-auto whitespace-pre-wrap break-words rounded-none bg-foreground p-3 font-mono text-xs leading-6 text-background">
-        {formatJson(toolCall.arguments)}
-      </pre>
-    </details>
-  );
-}
-
-function MetaItem({ label, value }: { label: string; value: string }) {
-  return (
-    <div className="space-y-1">
-      <p className="text-xs uppercase tracking-wide text-muted-foreground">{label}</p>
-      <p className="break-words text-sm font-medium">{value}</p>
     </div>
   );
 }
